server/requirements.txt
Django>=3.1,<3.2
django-cors-headers>=3.7,<3.8
django-rest-framework